+++ src/Sutil/Html.fs
@@ -8,8 +8,10 @@ open System
 // Dummy type to avoid problems with overload resolution in HtmlEngine
 type [<Fable.Core.Erase>] NodeAttr = NodeAttr of NodeFactory
 
-type SutilHtmlEngine(helper) =
-    inherit HtmlEngine<NodeFactory>(helper)
+type SutilHtmlEngine() =
+    inherit HtmlEngine<NodeFactory>(makeNode     = Func<string,NodeFactory seq,NodeFactory>(el),
+                                    stringToNode = Func<string,NodeFactory>(text),
+                                    emptyNode    = Func<NodeFactory>(fun () -> fragment []))
     member _.app (xs : seq<NodeFactory>) : NodeFactory = fragment xs
     member _.body (xs: seq<NodeFactory>) = nodeFactory <| fun ctx ->
         ctx |> ContextHelpers.withParent (ctx.Document.body) |> buildChildren xs
@@ -23,8 +25,10 @@ type SutilHtmlEngine(helper) =
 
     member _.fragment (v : IObservable<NodeFactory>) = Bind.fragment v id
 
-type SutilAttrEngine(helper) =
-    inherit AttrEngine<NodeFactory>(helper)
+type SutilAttrEngine() =
+    inherit AttrEngine<NodeFactory>(
+                makeAttr        = Func<string,string,NodeFactory>(fun key value -> attr(key, value)),
+                makeBooleanAttr = Func<string,bool,NodeFactory>(fun key value -> attr(key, value)))
 
     member _.disabled<'T> (value: IObservable<'T>) = bindAttrIn "disabled" value
 
@@ -37,17 +41,9 @@ type SutilAttrEngine(helper) =
     member _.value<'T> (value: IObservable<'T>, dispatch: 'T -> unit) =
             bindAttrBoth "value" value dispatch
 
-let Html = SutilHtmlEngine( {new HtmlHelper<NodeFactory> with
-            member _.MakeNode(tag, nodes) = el tag nodes
-            member _.StringToNode(v) = text v
-            member _.EmptyNode = fragment []
-            })
+let Html = SutilHtmlEngine()
 
-let Attr =
-    SutilAttrEngine
-        { new AttrHelper<NodeFactory> with
-            member _.MakeAttr(key : string, value : string) = attr(key, value)
-            member _.MakeBooleanAttr(key, value) = attr(key, value) }
+let Attr = SutilAttrEngine()
 
 // Convenience
 let text s = DOM.text s